+++ tsconfig.json
@@ -2,7 +2,8 @@
     "exclude": [                                /* Excluded folders from compilation */
       "node_modules",
       "dist",
-      "test",                                   /* Tests are ran with ts-jest, which processes TS files directly */
+      "test/*",                                   /* Tests are run with ts-jest, which processes TS files directly */
+      "src/**/*.test.*"
     ],
     "compilerOptions": {
       /* Visit https://aka.ms/tsconfig.json to read more about this file */
